Het ki 2

+++ Java/School_Exercise/Java_SangT3/src/Thi_Sang_T3/QuanLyHH.java
@@ -88,7 +88,7 @@ public class QuanLyHH {
 				}
 			}
 		}
-		return count;
+		return (int)count;
 	}
 	
 	public void xoaOto() {
